Split out files for extension. Template stuff not quite working yet
